v1.1/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
mysql-connector-python==8.0.33
protobuf==3.20.3
requests==2.30.0
soupsieve==2.4.1
termcolor==2.3.0
urllib3==2.0.2
